Move process args out of data, to avoid keras import and error/warning logging
