+++ zea/data/process.py
@@ -7,9 +7,8 @@ Usage:
 import argparse
 import re
 from concurrent.futures import ThreadPoolExecutor
-from dataclasses import dataclass, field, fields as dataclass_fields
+from dataclasses import fields as dataclass_fields
 from pathlib import Path
-from typing import Annotated
 
 import keras
 import numpy as np
@@ -17,6 +16,7 @@ import tyro
 from keras import ops
 
 from zea import display, io_lib, log
+from zea.cli_args import ProcessArgs
 from zea.config import Config
 from zea.data.dataloader import Dataloader
 from zea.data.datasets import Dataset
@@ -34,46 +34,6 @@ try:
     SUPPORTED_FORMATS += ["nii.gz"]
 except ImportError:
     sitk = None
-
-
-@dataclass
-class ProcessArgs:
-    """Arguments for beamforming a zea dataset."""
-
-    dataset: Annotated[
-        str,
-        tyro.conf.arg(
-            aliases=["-d"],
-            help="Path/URI to the zea dataset (folder of HDF5 files or a single HDF5 file).",
-        ),
-    ]
-    config: Annotated[
-        str,
-        tyro.conf.arg(
-            aliases=["-c"],
-            help="Path to config.yaml for the beamforming pipeline.",
-        ),
-    ]
-    save_dir: Path = Path("output")
-    key: str = "data/raw_data"
-    n_frames: int | None = None
-    save_as: str = "gif"
-    keep_keys: list[str] = field(default_factory=lambda: ["maxval"])
-    timings: bool = False
-    num_threads: int = 16
-    revision: str | None = None
-    config_revision: str | None = None
-    overwrite: bool = False
-    keep_dynamic_range: bool = False
-    device: Annotated[
-        str,
-        tyro.conf.arg(
-            help=(
-                "Compute device ('cuda:0', 'cpu', 'auto:1', …). "
-                "Only relevant when running the beamformer pipeline."
-            ),
-        ),
-    ] = "auto:1"
 
 
 def get_parser(add_help: bool = True) -> argparse.ArgumentParser:
